% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labo3.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm is an algorithm for finding the shortest paths between nodes in a graph, which may represent, for example, road networks.}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{graph}

\item{init_node}{starting node}
}
\value{
it return the shortest node every node from the initial node
}
\description{
Dijkstra's algorithm is an algorithm for finding the shortest paths between nodes in a graph, which may represent, for example, road networks.
}
\examples{
wiki_graph <-
data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),
           v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),
           w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
