% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labo3.R
\name{euclidean}
\alias{euclidean}
\title{This function is ued to find the greatest common devisier}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{first value}

\item{y}{second value}
}
\value{
it return the greate common devisier
}
\description{
This function is ued to find the greatest common devisier
}
\examples{
euclidean(123612,13892347912)
euclidean(100,1000)
}
